Use USM instead of Buffer
