+++ include/oneapi/dpl/pstl/hetero/dpcpp/parallel_backend_esimd_radix_sort.h
@@ -18,936 +18,17 @@
 
 #include <ext/intel/esimd.hpp>
 #include "sycl_defs.h"
-#include "esimd_radix_sort_utils.h"
+
+#include "esimd_radix_sort/esimd_radix_sort_one_wg.h"
+#include "esimd_radix_sort/esimd_radix_sort_cooperative.h"
+#include "esimd_radix_sort/esimd_radix_sort_onesweep.h"
+
+#include "../../utils.h"
+#include "../../utils_ranges.h"
+#include "utils_ranges_sycl.h"
 
 #include <cstdint>
-
-namespace oneapi::dpl::experimental::esimd::impl
-{
-template <typename Acc, typename KeyT, uint32_t RADIX_BITS, uint32_t PROCESS_SIZE>
-void
-one_workgroup_kernel(const Acc& acc, auto idx, uint32_t n, uint32_t THREAD_PER_TG)
-{
-    using namespace sycl;
-    using namespace __ESIMD_NS;
-    using namespace __ESIMD_ENS;
-
-    using bin_t = uint16_t;
-    using hist_t = uint16_t;
-    using device_addr_t = uint32_t;
-
-    uint32_t local_tid = idx.get_local_linear_id();
-    constexpr uint32_t BIN_COUNT = 1 << RADIX_BITS;
-    constexpr uint32_t NBITS = sizeof(KeyT) * 8;
-    constexpr uint32_t STAGES = div_up(NBITS, RADIX_BITS);
-    constexpr uint32_t MAX_THREAD_PER_TG = 64;
-    constexpr bin_t MASK = BIN_COUNT - 1;
-
-    constexpr uint32_t REORDER_SLM_SIZE = PROCESS_SIZE * sizeof(KeyT) * MAX_THREAD_PER_TG; // reorder buffer
-    constexpr uint32_t BIN_HIST_SLM_SIZE = BIN_COUNT * sizeof(hist_t) * MAX_THREAD_PER_TG; // bin hist working buffer
-    constexpr uint32_t INCOMING_OFFSET_SLM_SIZE = (BIN_COUNT + 1) * sizeof(hist_t);        // incoming offset buffer
-
-    // max SLM is 256 * 4 * 64 + 256 * 2 * 64 + 257*2, 97KB, when  PROCESS_SIZE = 256, BIN_COUNT = 256
-    // to support 512 processing size, we can use all SLM as reorder buffer with cost of more barrier
-    slm_init(std::max(REORDER_SLM_SIZE, BIN_HIST_SLM_SIZE + INCOMING_OFFSET_SLM_SIZE));
-    uint32_t slm_reorder_start = 0;
-    uint32_t slm_bin_hist_start = 0;
-    uint32_t slm_incoming_offset = slm_bin_hist_start + BIN_HIST_SLM_SIZE;
-
-    uint32_t slm_reorder = slm_reorder_start + local_tid * PROCESS_SIZE * sizeof(KeyT);
-    uint32_t slm_bin_hist_this_thread = slm_bin_hist_start + local_tid * BIN_COUNT * sizeof(hist_t);
-
-    simd<hist_t, BIN_COUNT> bin_offset;
-    simd<device_addr_t, PROCESS_SIZE> write_addr;
-    simd<KeyT, PROCESS_SIZE> keys;
-    simd<bin_t, PROCESS_SIZE> bins;
-    simd<device_addr_t, 16> lane_id(0, 1);
-
-    device_addr_t io_offset = PROCESS_SIZE * local_tid;
-
-#pragma unroll
-    for (uint32_t s = 0; s < PROCESS_SIZE; s += 16)
-    {
-        simd_mask<16> m = (io_offset + lane_id + s) < n;
-        keys.template select<16, 1>(s) =
-            // merge(gather(p_acc + io_offset + s, lane_id * uint32_t(sizeof(KeyT))),
-            //       simd<KeyT, 16>(-1), m);
-            merge(gather<KeyT>(acc, lane_id * uint32_t(sizeof(KeyT)), (io_offset + s) * sizeof(KeyT)),
-                  simd<KeyT, 16>(-1), m);
-    }
-
-    for (uint32_t stage = 0; stage < STAGES; stage++)
-    {
-        bins = (keys >> (stage * RADIX_BITS)) & MASK;
-        bin_offset = 0;
-#pragma unroll
-        for (uint32_t s = 0; s < PROCESS_SIZE; s += 1)
-        {
-            write_addr[s] = bin_offset[bins[s]];
-            bin_offset[bins[s]] += 1;
-        }
-
-            /*
-            first write to slm,
-            then do column scan by group, each thread to 32c*8r,
-            then last row do exclusive scan as incoming offset
-            then every thread add local sum with sum of previous group and incoming offset
-            */
-        {
-            barrier();
-            constexpr uint32_t HIST_STRIDE = sizeof(hist_t) * BIN_COUNT;
-#pragma unroll
-            for (uint32_t s = 0; s < BIN_COUNT; s += 128)
-            {
-                lsc_slm_block_store<uint32_t, 64>(
-                    slm_bin_hist_this_thread + s * sizeof(hist_t),
-                    bin_offset.template select<128, 1>(s).template bit_cast_view<uint32_t>());
-            }
-            barrier();
-            constexpr uint32_t BIN_SUMMARY_GROUP_SIZE = 8;
-            if (local_tid < BIN_SUMMARY_GROUP_SIZE)
-            {
-                constexpr uint32_t BIN_WIDTH = BIN_COUNT / BIN_SUMMARY_GROUP_SIZE;
-                constexpr uint32_t BIN_WIDTH_UD = BIN_WIDTH * sizeof(hist_t) / sizeof(uint32_t);
-                uint32_t slm_bin_hist_summary_offset = slm_bin_hist_start + local_tid * BIN_WIDTH * sizeof(hist_t);
-                simd<hist_t, BIN_WIDTH> thread_grf_hist_summary;
-                simd<uint32_t, BIN_WIDTH_UD> tmp;
-
-                thread_grf_hist_summary.template bit_cast_view<uint32_t>() =
-                    lsc_slm_block_load<uint32_t, BIN_WIDTH_UD>(slm_bin_hist_summary_offset);
-                slm_bin_hist_summary_offset += HIST_STRIDE;
-                for (uint32_t s = 1; s < THREAD_PER_TG - 1; s++)
-                {
-                    tmp = lsc_slm_block_load<uint32_t, BIN_WIDTH_UD>(slm_bin_hist_summary_offset);
-                    thread_grf_hist_summary += tmp.template bit_cast_view<hist_t>();
-                    lsc_slm_block_store<uint32_t, BIN_WIDTH_UD>(
-                        slm_bin_hist_summary_offset, thread_grf_hist_summary.template bit_cast_view<uint32_t>());
-                    slm_bin_hist_summary_offset += HIST_STRIDE;
-                }
-                tmp = lsc_slm_block_load<uint32_t, BIN_WIDTH_UD>(slm_bin_hist_summary_offset);
-                thread_grf_hist_summary += tmp.template bit_cast_view<hist_t>();
-                thread_grf_hist_summary = scan<hist_t, hist_t>(thread_grf_hist_summary);
-                lsc_slm_block_store<uint32_t, BIN_WIDTH_UD>(slm_bin_hist_summary_offset,
-                                                            thread_grf_hist_summary.template bit_cast_view<uint32_t>());
-            }
-            barrier();
-            if (local_tid == 0)
-            {
-                simd<hist_t, BIN_COUNT> grf_hist_summary;
-                simd<hist_t, BIN_COUNT + 1> grf_hist_summary_scan;
-#pragma unroll
-                for (uint32_t s = 0; s < BIN_COUNT; s += 128)
-                {
-                    grf_hist_summary.template select<128, 1>(s).template bit_cast_view<uint32_t>() =
-                        lsc_slm_block_load<uint32_t, 64>(slm_bin_hist_start + (THREAD_PER_TG - 1) * HIST_STRIDE +
-                                                         s * sizeof(hist_t));
-                }
-                grf_hist_summary_scan[0] = 0;
-                grf_hist_summary_scan.template select<32, 1>(1) = grf_hist_summary.template select<32, 1>(0);
-#pragma unroll
-                for (uint32_t i = 32; i < BIN_COUNT; i += 32)
-                {
-                    grf_hist_summary_scan.template select<32, 1>(i + 1) =
-                        grf_hist_summary.template select<32, 1>(i) + grf_hist_summary_scan[i];
-                }
-#pragma unroll
-                for (uint32_t s = 0; s < BIN_COUNT; s += 128)
-                {
-                    lsc_slm_block_store<uint32_t, 64>(
-                        slm_incoming_offset + s * sizeof(hist_t),
-                        grf_hist_summary_scan.template select<128, 1>(s).template bit_cast_view<uint32_t>());
-                }
-            }
-            barrier();
-            {
-#pragma unroll
-                for (uint32_t s = 0; s < BIN_COUNT; s += 128)
-                {
-                    bin_offset.template select<128, 1>(s).template bit_cast_view<uint32_t>() =
-                        lsc_slm_block_load<uint32_t, 64>(slm_incoming_offset + s * sizeof(hist_t));
-                }
-                if (local_tid > 0)
-                {
-#pragma unroll
-                    for (uint32_t s = 0; s < BIN_COUNT; s += 128)
-                    {
-                        simd<hist_t, 128> group_local_sum;
-                        group_local_sum.template bit_cast_view<uint32_t>() = lsc_slm_block_load<uint32_t, 64>(
-                            slm_bin_hist_start + (local_tid - 1) * HIST_STRIDE + s * sizeof(hist_t));
-                        bin_offset.template select<128, 1>(s) += group_local_sum;
-                    }
-                }
-            }
-            barrier();
-        }
-
-#pragma unroll
-        for (uint32_t s = 0; s < PROCESS_SIZE; s += 16)
-        {
-            simd<uint16_t, 16> bins_uw = bins.template select<16, 1>(s);
-            write_addr.template select<16, 1>(s) += bin_offset.template iselect(bins_uw);
-        }
-
-        if (stage != STAGES - 1)
-        {
-#pragma unroll
-            for (uint32_t s = 0; s < PROCESS_SIZE; s += 16)
-            {
-                slm_scatter<KeyT, 16>(write_addr.template select<16, 1>(s) * sizeof(KeyT) +
-                                                      slm_reorder_start,
-                                                  keys.template select<16, 1>(s));
-            }
-            barrier();
-#pragma unroll
-            for (uint32_t s = 0; s < PROCESS_SIZE; s += 64)
-            {
-                keys.template select<64, 1>(s) =
-                    lsc_slm_block_load<KeyT, 64>(slm_reorder + s * sizeof(KeyT));
-            }
-        }
-    }
-#pragma unroll
-    for (uint32_t s = 0; s < PROCESS_SIZE; s += 16)
-    {
-        // scatter<KeyT, 16>(p_acc, write_addr.template select<16, 1>(s) * sizeof(KeyT),
-        //                               keys.template select<16, 1>(s), (local_tid * PROCESS_SIZE + lane_id + s) < n);
-        scatter<KeyT, 16>(acc, write_addr.template select<16, 1>(s) * sizeof(KeyT),
-                                      keys.template select<16, 1>(s), 0, (local_tid * PROCESS_SIZE + lane_id + s) < n);
-    }
-}
-
-template <typename _Range, typename KeyT, uint32_t RADIX_BITS>
-void
-one_workgroup(sycl::queue& q, _Range&& rng, std::size_t n)
-{
-    using namespace sycl;
-
-    constexpr uint32_t BINCOUNT = 1 << RADIX_BITS;
-    constexpr uint32_t MAX_TG_COUNT = 64;
-    constexpr uint32_t MIN_TG_COUNT = 8;
-    uint32_t PROCESS_SIZE = 64;
-    if (n < MIN_TG_COUNT * 64)
-    {
-        PROCESS_SIZE = 64;
-    }
-    else if (n < MIN_TG_COUNT * 128)
-    {
-        PROCESS_SIZE = 128;
-    }
-    else
-    {
-        PROCESS_SIZE = 256;
-    };
-    assert((PROCESS_SIZE == 64) || (PROCESS_SIZE == 128) || (PROCESS_SIZE == 256));
-
-    uint32_t TG_COUNT = div_up(n, PROCESS_SIZE);
-    TG_COUNT = std::max(TG_COUNT, MIN_TG_COUNT);
-    assert((TG_COUNT <= MAX_TG_COUNT) && (TG_COUNT >= MIN_TG_COUNT));
-
-    auto acc = rng.__m_acc;
-    sycl::event e;
-    {
-        nd_range<1> Range((range<1>(TG_COUNT)), (range<1>(TG_COUNT)));
-        if (PROCESS_SIZE == 64)
-        {
-            e = q.submit([&](handler& cgh) {
-                oneapi::dpl::__ranges::__require_access(cgh, rng);
-                cgh.parallel_for<class kernel_radix_sort_one_tg_64>(Range, [=](nd_item<1> idx) [[intel::sycl_explicit_simd]] {
-                    one_workgroup_kernel<decltype(acc), KeyT, RADIX_BITS, 64>(acc, idx, n, TG_COUNT);
-                });
-            });
-        }
-        else if (PROCESS_SIZE == 128)
-        {
-            e = q.submit([&](handler& cgh) {
-                oneapi::dpl::__ranges::__require_access(cgh, rng);
-                cgh.parallel_for<class kernel_radix_sort_one_tg_128>(Range, [=](nd_item<1> idx) [[intel::sycl_explicit_simd]] {
-                    one_workgroup_kernel<decltype(acc), KeyT, RADIX_BITS, 128>(acc, idx, n, TG_COUNT);
-                });
-            });
-        }
-        else if (PROCESS_SIZE == 256)
-        {
-            e = q.submit([&](handler& cgh) {
-                oneapi::dpl::__ranges::__require_access(cgh, rng);
-                cgh.parallel_for<class kernel_radix_sort_one_tg_256>(Range, [=](nd_item<1> idx) [[intel::sycl_explicit_simd]] {
-                    one_workgroup_kernel<decltype(acc), KeyT, RADIX_BITS, 256>(acc, idx, n, TG_COUNT);
-                });
-            });
-        }
-    }
-    e.wait();
-}
-
-template <typename Acc, typename KeyT, uint32_t RADIX_BITS, uint32_t THREAD_PER_TG, uint32_t PROCESS_SIZE>
-void cooperative_kernel(const Acc& acc, auto idx, size_t n, uint32_t *p_global_buffer) {
-    using namespace sycl;
-    using namespace __ESIMD_NS;
-    using namespace __ESIMD_ENS;
-
-    using bin_t = uint16_t;
-    using hist_t = uint32_t;
-    using global_hist_t = uint32_t;
-    using device_addr_t = uint32_t;
-
-    uint32_t tg_id = idx.get_group(0);
-    uint32_t tg_count = idx.get_group_range(0);
-
-    uint32_t local_tid = idx.get_local_linear_id();
-    constexpr uint32_t BIN_COUNT = 1 << RADIX_BITS;
-    constexpr uint32_t NBITS =  sizeof(KeyT) * 8;
-    constexpr uint32_t STAGES = div_up(NBITS, RADIX_BITS);
-    constexpr uint32_t TG_PROCESS_SIZE = PROCESS_SIZE * THREAD_PER_TG;
-    constexpr bin_t MASK = BIN_COUNT - 1;
-
-    constexpr uint32_t BIN_HIST_SLM_SIZE = BIN_COUNT * sizeof(hist_t) * THREAD_PER_TG;  // bin hist working buffer, 64K for DW hist
-    constexpr uint32_t INCOMING_OFFSET_SLM_SIZE = (BIN_COUNT+16)*sizeof(global_hist_t); // incoming offset buffer
-    constexpr uint32_t GLOBAL_SCAN_SIZE = (BIN_COUNT+16)*sizeof(global_hist_t);
-
-    slm_init( BIN_HIST_SLM_SIZE + INCOMING_OFFSET_SLM_SIZE + GLOBAL_SCAN_SIZE);
-    uint32_t slm_bin_hist_start = 0;
-    uint32_t slm_incoming_start = slm_bin_hist_start + BIN_HIST_SLM_SIZE;
-    uint32_t slm_global_scan_start = slm_incoming_start + INCOMING_OFFSET_SLM_SIZE;
-
-    uint32_t slm_bin_hist_this_thread = slm_bin_hist_start + local_tid * BIN_COUNT * sizeof(hist_t);
-
-    uint32_t global_sync_buffer_size = 1024; //1K uint32_t for sync buffer
-    uint32_t global_bin_start_buffer_size = BIN_COUNT * sizeof(global_hist_t)+16;
-    uint32_t global_bin_hist_size = tg_count * BIN_COUNT * sizeof(global_hist_t);
-    uint32_t *p_sync_buffer = p_global_buffer;
-    uint32_t *p_global_bin_start_buffer = p_sync_buffer + global_sync_buffer_size;
-    uint32_t *p_global_bin_hist = p_global_bin_start_buffer + global_bin_start_buffer_size;
-    uint32_t *p_global_bin_hist_tg = p_global_bin_hist + tg_id * BIN_COUNT * sizeof(global_hist_t)/sizeof(uint32_t);
-
-    simd<hist_t, BIN_COUNT> bin_offset;
-    simd<device_addr_t, PROCESS_SIZE> write_addr;
-    simd<KeyT, PROCESS_SIZE> keys;
-    simd<bin_t, PROCESS_SIZE> bins;
-    simd<device_addr_t, 16> lane_id(0, 1);
-
-    device_addr_t io_offset = tg_id * TG_PROCESS_SIZE + PROCESS_SIZE * local_tid;
-
-    constexpr uint32_t BIN_GROUPS = 8;
-    constexpr uint32_t THREAD_PER_BIN_GROUP = THREAD_PER_TG / BIN_GROUPS;
-    constexpr uint32_t BIN_WIDTH = BIN_COUNT / BIN_GROUPS;
-    constexpr uint32_t BIN_WIDTH_UD = BIN_COUNT / BIN_GROUPS * sizeof(hist_t) / sizeof(uint32_t);
-    constexpr uint32_t BIN_HEIGHT = THREAD_PER_TG / THREAD_PER_BIN_GROUP;
-
-    if (local_tid == 0) init_global_sync(p_sync_buffer, tg_id);
-    barrier();
-
-    #pragma unroll
-    for (uint32_t s = 0; s<PROCESS_SIZE; s+=16) {
-        simd_mask<16> m = (io_offset+lane_id+s)<n;
-        simd<KeyT, 16> source = lsc_gather<KeyT, 1,
-                // lsc_data_size::default_size, cache_hint::uncached, cache_hint::cached, 16>(acc+io_offset+s, lane_id*uint32_t(sizeof(KeyT)), m);
-                lsc_data_size::default_size, cache_hint::uncached, cache_hint::cached, 16>(acc, sycl::ext::intel::esimd::simd<KeyT, 16>((lane_id + io_offset+s)*uint32_t(sizeof(KeyT))), m);
-        keys.template select<16, 1>(s) = merge(source, simd<KeyT, 16>(-1), m);
-    }
-
-    for (uint32_t stage=0; stage < STAGES; stage++) {
-        bins = (keys >> (stage * RADIX_BITS)) & MASK;
-        bin_offset = 0;
-        #pragma unroll
-        for (uint32_t s = 0; s<PROCESS_SIZE; s+=1) {
-            write_addr[s] = bin_offset[bins[s]];
-            bin_offset[bins[s]] += 1;
-        }
-        /*
-        first write to slm,
-        then do column scan by group, each thread to 32c*8r,
-        then last row do exclusive scan as incoming offset
-        then every thread add local sum with sum of previous group and incoming offset
-        */
-        {
-            barrier();
-            // put local hist to slm
-            #pragma unroll
-            for (uint32_t s = 0; s<BIN_COUNT; s+=64) {
-                lsc_slm_block_store<uint32_t, 64>(slm_bin_hist_this_thread + s*sizeof(hist_t), bin_offset.template select<64, 1>(s).template bit_cast_view<uint32_t>());
-            }
-            barrier();
-            // small group sum for local hist
-            constexpr uint32_t BIN_GROUPS = 8;
-            constexpr uint32_t THREAD_PER_BIN_GROUP = THREAD_PER_TG / BIN_GROUPS;
-            constexpr uint32_t BIN_WIDTH = BIN_COUNT / BIN_GROUPS;
-            constexpr uint32_t BIN_HEIGHT = THREAD_PER_TG / THREAD_PER_BIN_GROUP;
-            constexpr uint32_t HIST_STRIDE = BIN_COUNT * sizeof(hist_t);
-            uint32_t THREAD_GID = local_tid / THREAD_PER_BIN_GROUP;
-            uint32_t THREAD_LTID = local_tid % THREAD_PER_BIN_GROUP;
-            {
-                uint32_t slm_bin_hist_ingroup_offset = slm_bin_hist_start + THREAD_GID * BIN_WIDTH * sizeof(hist_t) + THREAD_LTID * BIN_HEIGHT * HIST_STRIDE;
-                simd2d<hist_t, BIN_HEIGHT, BIN_WIDTH> thread_grf_hist;
-                #pragma unroll
-                for (uint32_t s = 0; s<BIN_HEIGHT; s++) {
-                    thread_grf_hist.row(s).template bit_cast_view<uint32_t>() = lsc_slm_block_load<uint32_t, BIN_WIDTH_UD>(slm_bin_hist_ingroup_offset + s * HIST_STRIDE);
-                }
-                #pragma unroll
-                for (uint32_t s = 1; s<BIN_HEIGHT; s++) {
-                    thread_grf_hist.row(s) += thread_grf_hist.row(s-1);
-                }
-                #pragma unroll
-                for (uint32_t s = 1; s<BIN_HEIGHT; s++) {
-                    lsc_slm_block_store<uint32_t, BIN_WIDTH_UD>(slm_bin_hist_ingroup_offset + s * HIST_STRIDE, thread_grf_hist.row(s).template bit_cast_view<uint32_t>());
-                }
-            }
-
-            barrier();
-            // thread group sum for groups
-            if (THREAD_LTID == 0) {
-                uint32_t slm_bin_hist_group_summary_offset = slm_bin_hist_start + THREAD_GID * BIN_WIDTH * sizeof(hist_t) +  (BIN_HEIGHT-1) * HIST_STRIDE;
-                simd2d<hist_t, THREAD_PER_BIN_GROUP, BIN_WIDTH> thread_grf_hist_summary;
-                #pragma unroll
-                for (uint32_t s = 0; s<THREAD_PER_BIN_GROUP; s++) {
-                    thread_grf_hist_summary.row(s).template bit_cast_view<uint32_t>() = lsc_slm_block_load<uint32_t, BIN_WIDTH_UD>(slm_bin_hist_group_summary_offset + s * BIN_HEIGHT * HIST_STRIDE);
-                }
-                #pragma unroll
-                for (uint32_t s = 1; s<THREAD_PER_BIN_GROUP; s++) {
-                    thread_grf_hist_summary.row(s) += thread_grf_hist_summary.row(s-1);
-                }
-                #pragma unroll
-                for (uint32_t s = 1; s<THREAD_PER_BIN_GROUP-1; s++) {
-                    lsc_slm_block_store<uint32_t, BIN_WIDTH_UD>(slm_bin_hist_group_summary_offset + s * BIN_HEIGHT * HIST_STRIDE, thread_grf_hist_summary.row(s).template bit_cast_view<uint32_t>());
-                }
-                simd<global_hist_t, BIN_WIDTH> group_hist_sum = thread_grf_hist_summary.row(THREAD_PER_BIN_GROUP-1);
-                group_hist_sum.copy_to(p_global_bin_hist_tg + THREAD_GID * BIN_WIDTH * sizeof(global_hist_t)/sizeof(uint32_t));
-            }
-            lsc_fence<lsc_memory_kind::untyped_global, lsc_fence_op::evict, lsc_scope::gpu>();
-            barrier();
-
-            // thread group scan -> change to coopeative group scan
-            // first dump result to global hist buffer - done in previous part, last row is copy to global
-            // global sync()
-            // each tg calculate sum of previous tg
-            // last tg then do exclusive scan and save to global scan buffer
-            // global sync()
-            // each tg read global scan buffer, add with prev tg, then put to slm incoming buffer
-            if (local_tid == 0) {
-                global_sync(p_sync_buffer, stage*4+0, tg_count, tg_id, local_tid);
-            }
-            barrier();
-            {
-                if (tg_id == tg_count-1) {
-                    if (local_tid < 16)  {
-                        //16 threads cooperative, each for BIN_COUNT/16 element.
-                        uint32_t *p_global_hist_sum_per_thread = p_global_bin_hist + local_tid * (BIN_COUNT/16)* sizeof(global_hist_t)/sizeof(uint32_t);
-
-                        simd<global_hist_t, BIN_COUNT/16> global_hist_sum(p_global_hist_sum_per_thread);
-                        for (uint32_t i = 1; i<tg_count; i++) {
-                            simd<global_hist_t, BIN_COUNT/16> global_hist(p_global_hist_sum_per_thread + i * BIN_COUNT * sizeof(global_hist_t)/sizeof(uint32_t));
-                            global_hist_sum += global_hist;
-                            global_hist_sum.copy_to(p_global_hist_sum_per_thread + i * BIN_COUNT * sizeof(global_hist_t)/sizeof(uint32_t));
-                        }
-                        global_hist_sum = scan<global_hist_t, global_hist_t>(global_hist_sum);
-                        lsc_slm_block_store<global_hist_t, 16>(slm_global_scan_start + local_tid * (BIN_COUNT/16)* sizeof(global_hist_t), global_hist_sum);
-                    }
-                    barrier();//TODO: change to smaller named barrier
-                    if (local_tid==0) {
-                        simd<global_hist_t, BIN_COUNT/16> global_hist_sum;
-                        global_hist_t prev(0);
-                        #pragma unroll
-                        for (uint32_t i = 0; i<BIN_COUNT; i+=16) {
-                            global_hist_sum = prev + lsc_slm_block_load<global_hist_t, 16>(slm_global_scan_start + i * sizeof(global_hist_t));
-                            prev = global_hist_sum[15];
-                            global_hist_sum.copy_to(p_global_bin_start_buffer+1+i);
-                        }
-                        p_global_bin_start_buffer[0] = 0;
-                    }
-                    lsc_fence<lsc_memory_kind::untyped_global, lsc_fence_op::evict, lsc_scope::gpu>();
-                }
-                // if (stage == 0) return;
-                if (local_tid == 0)  {
-                    global_sync(p_sync_buffer, stage*4+1, tg_count, tg_id, local_tid);
-                    {
-                        simd<global_hist_t, BIN_COUNT> global_hist_start(p_global_bin_start_buffer);
-                        if (tg_id != 0) {
-                            global_hist_start += simd<global_hist_t, BIN_COUNT>(p_global_bin_hist_tg - BIN_COUNT * sizeof(global_hist_t)/sizeof(uint32_t));
-                        }
-                        #pragma unroll
-                        for (uint32_t s = 0; s<BIN_COUNT; s+=64) {
-                            lsc_slm_block_store<uint32_t, 64>(slm_incoming_start + s * sizeof(global_hist_t), global_hist_start.template select<64, 1>(s));
-                        }
-                    }
-                }
-            }
-            barrier();
-            {
-                #pragma unroll
-                for (uint32_t s = 0; s<BIN_COUNT; s+=64) {
-                    bin_offset.template select<64, 1>(s).template bit_cast_view<uint32_t>() = lsc_slm_block_load<uint32_t, 64>(slm_incoming_start + s*sizeof(hist_t));
-                }
-                if (local_tid>0) {
-                    #pragma unroll
-                    for (uint32_t s = 0; s<BIN_COUNT; s+=64) {
-                        simd<hist_t, 64> group_local_sum;
-                        group_local_sum.template bit_cast_view<uint32_t>() = lsc_slm_block_load<uint32_t, 64>(slm_bin_hist_start + (local_tid-1)*HIST_STRIDE + s*sizeof(hist_t));
-                        bin_offset.template select<64, 1>(s) += group_local_sum;
-                    }
-                }
-                if ((local_tid > THREAD_PER_BIN_GROUP) && (local_tid % BIN_HEIGHT != 0)) {
-                    uint32_t prev_cum_rowid = (local_tid-1) / BIN_HEIGHT * BIN_HEIGHT - 1;
-                    #pragma unroll
-                    for (uint32_t s = 0; s<BIN_COUNT; s+=64) {
-                        simd<hist_t, 64> group_sum;
-                        group_sum.template bit_cast_view<uint32_t>() = lsc_slm_block_load<uint32_t, 64>(slm_bin_hist_start + prev_cum_rowid*HIST_STRIDE + s*sizeof(hist_t));
-                        bin_offset.template select<64, 1>(s) += group_sum;
-                    }
-                }
-            }
-            barrier();
-        }
-
-        #pragma unroll
-        for (uint32_t s = 0; s<PROCESS_SIZE; s+=16) {
-            simd<uint16_t, 16> bins_uw = bins.template select<16, 1>(s);
-            write_addr.template select<16, 1>(s) += bin_offset.template iselect(bins_uw);
-        }
-
-        #pragma unroll
-        for (uint32_t s = 0; s<PROCESS_SIZE; s+=16) {
-            lsc_scatter<KeyT, 1, lsc_data_size::default_size, cache_hint::uncached, cache_hint::write_back, 16>(
-                acc,
-                write_addr.template select<16, 1>(s)*sizeof(KeyT),
-                keys.template select<16, 1>(s), write_addr.template select<16, 1>(s)<n);
-        }
-        if (stage != STAGES - 1) {
-            lsc_fence<lsc_memory_kind::untyped_global, lsc_fence_op::evict, lsc_scope::gpu>();
-            barrier();
-            if (local_tid == 0) {global_sync(p_sync_buffer, stage*4+2, tg_count, tg_id, local_tid);}
-            barrier();
-            #pragma unroll
-            for (uint32_t s = 0; s<PROCESS_SIZE; s+=16) {
-                simd_mask<16> m = (io_offset+lane_id+s)<n;
-                simd<KeyT, 16> reordered = lsc_gather<KeyT, 1,
-                        lsc_data_size::default_size, cache_hint::uncached, cache_hint::cached, 16>(
-                            // acc+io_offset+s, lane_id*uint32_t(sizeof(KeyT)), m);
-                            acc, sycl::ext::intel::esimd::simd<KeyT, 16>((lane_id + io_offset+s)*uint32_t(sizeof(KeyT))), m);
-                keys.template select<16, 1>(s) = merge(reordered, simd<KeyT, 16>(-1), m);
-            }
-        }
-    }
-}
-
-template <typename _Range, typename KeyT, uint32_t RADIX_BITS>
-void cooperative(sycl::queue &q, _Range&& rng, std::size_t n) {
-    using namespace sycl;
-    using namespace __ESIMD_NS;
-
-    constexpr uint32_t BIN_COUNT = 1 << RADIX_BITS;
-    constexpr uint32_t THREAD_PER_TG = 64;
-    uint32_t MAX_GROUPS = 56; //TODO: get from sycl api
-    assert(n <= 256 * THREAD_PER_TG * MAX_GROUPS);
-
-    auto acc = rng.__m_acc;
-    auto p_sync = static_cast<uint32_t*>(sycl::malloc_device(1024 + (MAX_GROUPS+2) * BIN_COUNT * sizeof(uint32_t), q));
-    sycl::event e;
-    {
-        if (n<=128 * THREAD_PER_TG * MAX_GROUPS) {
-            uint32_t groups = div_up(n, 128 * THREAD_PER_TG);
-            nd_range<1> Range( (range<1>(THREAD_PER_TG * groups)), (range<1>(THREAD_PER_TG)));
-            e = q.submit([&](handler &cgh) {
-                oneapi::dpl::__ranges::__require_access(cgh, rng);
-                cgh.parallel_for<class kernel_radix_sort_cooperative_128>(
-                        Range, [=](nd_item<1> idx) [[intel::sycl_explicit_simd]] {
-                            cooperative_kernel<decltype(acc), KeyT, RADIX_BITS, THREAD_PER_TG, 128> (acc, idx, n, p_sync);
-                        });
-            });
-        } else if (n<=256 * THREAD_PER_TG * MAX_GROUPS) {
-            uint32_t groups = div_up(n, 256 * THREAD_PER_TG);
-            nd_range<1> Range( (range<1>(THREAD_PER_TG * groups)), (range<1>(THREAD_PER_TG)));
-            e = q.submit([&](handler &cgh) {
-                oneapi::dpl::__ranges::__require_access(cgh, rng);
-                cgh.parallel_for<class kernel_radix_sort_cooperative_256>(
-                        Range, [=](nd_item<1> idx) [[intel::sycl_explicit_simd]] {
-                            cooperative_kernel<decltype(acc), KeyT, RADIX_BITS, THREAD_PER_TG, 256> (acc, idx, n, p_sync);
-                        });
-            });
-        }
-    }
-    e.wait();
-    sycl::free(p_sync, q);
-}
-
-template <typename InT, typename OutT, typename KeyT, uint32_t RADIX_BITS, uint32_t THREAD_PER_TG, uint32_t PROCESS_SIZE>
-void onesweep_kernel(auto idx, uint32_t n, uint32_t stage, InT in, OutT out, uint8_t *p_global_buffer) {
-    using namespace sycl;
-    using namespace __ESIMD_NS;
-    using namespace __ESIMD_ENS;
-
-    using bin_t = uint16_t;
-    using hist_t = uint32_t;
-    using global_hist_t = uint32_t;
-    using device_addr_t = uint32_t;
-
-    uint32_t tg_id = idx.get_group(0);
-    uint32_t tg_count = idx.get_group_range(0);
-
-    uint32_t local_tid = idx.get_local_linear_id();
-    constexpr uint32_t BIN_COUNT = 1 << RADIX_BITS;
-    constexpr uint32_t NBITS =  sizeof(KeyT) * 8;
-    constexpr uint32_t STAGES = div_up(NBITS, RADIX_BITS);
-    constexpr uint32_t TG_PROCESS_SIZE = PROCESS_SIZE * THREAD_PER_TG;
-    constexpr bin_t MASK = BIN_COUNT - 1;
-
-    constexpr uint32_t BIN_HIST_SLM_SIZE = BIN_COUNT * sizeof(hist_t) * THREAD_PER_TG;  // bin hist working buffer, 64K for DW hist
-    constexpr uint32_t INCOMING_OFFSET_SLM_SIZE = (BIN_COUNT+16)*sizeof(global_hist_t); // incoming offset buffer
-    constexpr uint32_t GLOBAL_SCAN_SIZE = (BIN_COUNT+16)*sizeof(global_hist_t);
-
-    slm_init( BIN_HIST_SLM_SIZE + INCOMING_OFFSET_SLM_SIZE + GLOBAL_SCAN_SIZE);
-    uint32_t slm_bin_hist_start = 0;
-    uint32_t slm_incoming_start = slm_bin_hist_start + BIN_HIST_SLM_SIZE;
-    uint32_t slm_global_scan_start = slm_incoming_start + INCOMING_OFFSET_SLM_SIZE;
-
-
-    uint32_t slm_bin_hist_this_thread = slm_bin_hist_start + local_tid * BIN_COUNT * sizeof(hist_t);
-
-    size_t temp_io_size = n * sizeof(KeyT);
-
-    global_hist_t *p_global_bin_start_buffer_allstages = reinterpret_cast<global_hist_t*>(p_global_buffer);
-    global_hist_t *p_global_bin_start_buffer = p_global_bin_start_buffer_allstages + BIN_COUNT * stage;
-    uint32_t *p_sync_buffer = reinterpret_cast<uint32_t*>(p_global_bin_start_buffer_allstages + BIN_COUNT * STAGES);
-
-    constexpr uint32_t SYNC_SEGMENT_COUNT = 64;
-    constexpr uint32_t SYNC_SEGMENT_SIZE_DW = 128;
-
-
-    simd<hist_t, BIN_COUNT> bin_offset;
-    simd<device_addr_t, 16> lane_id(0, 1);
-
-    device_addr_t io_offset = tg_id * TG_PROCESS_SIZE + PROCESS_SIZE * local_tid;
-
-    constexpr uint32_t BIN_GROUPS = 8;
-    constexpr uint32_t THREAD_PER_BIN_GROUP = THREAD_PER_TG / BIN_GROUPS;
-    constexpr uint32_t BIN_WIDTH = BIN_COUNT / BIN_GROUPS;
-    constexpr uint32_t BIN_WIDTH_UD = BIN_COUNT / BIN_GROUPS * sizeof(hist_t) / sizeof(uint32_t);
-    constexpr uint32_t BIN_HEIGHT = THREAD_PER_TG / THREAD_PER_BIN_GROUP;
-
-    {
-        simd<KeyT, PROCESS_SIZE> keys;
-        simd<bin_t, PROCESS_SIZE> bins;
-
-        if (io_offset+PROCESS_SIZE < n) {
-            // keys.copy_from(in+io_offset);
-            copy_from(keys, in, io_offset);
-        }
-        else if (io_offset >= n) {
-            keys = -1;
-        }
-        else {
-            #pragma unroll
-            for (uint32_t s = 0; s<PROCESS_SIZE; s+=16) {
-                simd_mask<16> m = (io_offset+lane_id+s)<n;
-                simd<KeyT, 16> source = lsc_gather<KeyT, 1,
-                        lsc_data_size::default_size, cache_hint::cached, cache_hint::cached, 16>(in, sycl::ext::intel::esimd::simd((lane_id + io_offset+s)*uint32_t(sizeof(KeyT))), m);
-                keys.template select<16, 1>(s) = merge(source, simd<KeyT, 16>(-1), m);
-            }
-        }
-        bins = (keys >> (stage * RADIX_BITS)) & MASK;
-        bin_offset = 0;
-
-        #pragma unroll
-        for (uint32_t s = 0; s<PROCESS_SIZE; s+=1) {
-            bin_offset[bins[s]] += 1;
-        }
-    }
-    {
-        /*
-        first write to slm,
-        then do column scan by group, each thread to 32c*8r,
-        then last row do exclusive scan as incoming offset
-        then every thread add local sum with sum of previous group and incoming offset
-        */
-        {
-            // put local hist to slm
-            #pragma unroll
-            for (uint32_t s = 0; s<BIN_COUNT; s+=64) {
-                lsc_slm_block_store<uint32_t, 64>(slm_bin_hist_this_thread + s*sizeof(hist_t), bin_offset.template select<64, 1>(s).template bit_cast_view<uint32_t>());
-            }
-            barrier();
-            // small group sum for local hist
-            constexpr uint32_t BIN_GROUPS = 8;
-            constexpr uint32_t THREAD_PER_BIN_GROUP = THREAD_PER_TG / BIN_GROUPS;
-            constexpr uint32_t BIN_WIDTH = BIN_COUNT / BIN_GROUPS;
-            constexpr uint32_t BIN_HEIGHT = THREAD_PER_TG / THREAD_PER_BIN_GROUP;
-            constexpr uint32_t HIST_STRIDE = BIN_COUNT * sizeof(hist_t);
-            uint32_t THREAD_GID = local_tid / THREAD_PER_BIN_GROUP;
-            uint32_t THREAD_LTID = local_tid % THREAD_PER_BIN_GROUP;
-            {
-                uint32_t slm_bin_hist_ingroup_offset = slm_bin_hist_start + THREAD_GID * BIN_WIDTH * sizeof(hist_t) + THREAD_LTID * BIN_HEIGHT * HIST_STRIDE;
-                simd2d<hist_t, BIN_HEIGHT, BIN_WIDTH> thread_grf_hist;
-                #pragma unroll
-                for (uint32_t s = 0; s<BIN_HEIGHT; s++) {
-                    thread_grf_hist.row(s).template bit_cast_view<uint32_t>() = lsc_slm_block_load<uint32_t, BIN_WIDTH_UD>(slm_bin_hist_ingroup_offset + s * HIST_STRIDE);
-                }
-                #pragma unroll
-                for (uint32_t s = 1; s<BIN_HEIGHT; s++) {
-                    thread_grf_hist.row(s) += thread_grf_hist.row(s-1);
-                }
-                #pragma unroll
-                for (uint32_t s = 1; s<BIN_HEIGHT; s++) {
-                    lsc_slm_block_store<uint32_t, BIN_WIDTH_UD>(slm_bin_hist_ingroup_offset + s * HIST_STRIDE, thread_grf_hist.row(s).template bit_cast_view<uint32_t>());
-                }
-            }
-
-            barrier();
-            // thread group sum for groups
-            simd<global_hist_t, BIN_WIDTH> group_hist_sum;
-            if (THREAD_LTID == 0) {
-                uint32_t slm_bin_hist_group_summary_offset = slm_bin_hist_start + THREAD_GID * BIN_WIDTH * sizeof(hist_t) +  (BIN_HEIGHT-1) * HIST_STRIDE;
-                simd2d<hist_t, THREAD_PER_BIN_GROUP, BIN_WIDTH> thread_grf_hist_summary;
-                #pragma unroll
-                for (uint32_t s = 0; s<THREAD_PER_BIN_GROUP; s++) {
-                    thread_grf_hist_summary.row(s).template bit_cast_view<uint32_t>() = lsc_slm_block_load<uint32_t, BIN_WIDTH_UD>(slm_bin_hist_group_summary_offset + s * BIN_HEIGHT * HIST_STRIDE);
-                }
-                #pragma unroll
-                for (uint32_t s = 1; s<THREAD_PER_BIN_GROUP; s++) {
-                    thread_grf_hist_summary.row(s) += thread_grf_hist_summary.row(s-1);
-                }
-                #pragma unroll
-                for (uint32_t s = 1; s<THREAD_PER_BIN_GROUP-1; s++) {
-                    lsc_slm_block_store<uint32_t, BIN_WIDTH_UD>(slm_bin_hist_group_summary_offset + s * BIN_HEIGHT * HIST_STRIDE, thread_grf_hist_summary.row(s).template bit_cast_view<uint32_t>());
-                }
-                group_hist_sum = thread_grf_hist_summary.row(THREAD_PER_BIN_GROUP-1);
-            }
-            // wait for previous TG pass over starting point
-            // atomic add starting point with group_hist_sum
-            // and pass to next
-            {
-                //sync offset will be distributed to 64 banks to reduce latency
-                //for each tg, sync segment id is (tg_id % 64), sync target is ((tg_id+63) / 64) * BIN_GROUPS,
-                if (local_tid == 0) {
-                    uint32_t sync_segment_id = tg_id % SYNC_SEGMENT_COUNT;
-                    uint32_t sync_segment_target = ((tg_id + SYNC_SEGMENT_COUNT-1) / SYNC_SEGMENT_COUNT) * BIN_GROUPS;
-                    uint32_t sync_offset = sync_segment_id * SYNC_SEGMENT_SIZE_DW + stage;
-                    global_wait(p_sync_buffer, sync_offset, sync_segment_target, tg_id, local_tid);
-                }
-                barrier();
-                if (THREAD_LTID == 0) {
-                    uint32_t pass_over_sync_segment_id = (tg_id + 1) % SYNC_SEGMENT_COUNT;
-                    uint32_t pass_over_sync_offset = pass_over_sync_segment_id * SYNC_SEGMENT_SIZE_DW + stage;
-                    simd<uint32_t, BIN_WIDTH> update_lanes(0, sizeof(global_hist_t));
-                    group_hist_sum = lsc_atomic_update<atomic_op::add, uint32_t, BIN_WIDTH>(p_global_bin_start_buffer+THREAD_GID*BIN_WIDTH, update_lanes, group_hist_sum, 1);
-                    lsc_slm_block_store<uint32_t, BIN_WIDTH>(slm_incoming_start + THREAD_GID * BIN_WIDTH * sizeof(global_hist_t), group_hist_sum);
-                    // lsc_atomic_update<atomic_op::inc, uint32_t, 1>(p_sync_buffer+pass_over_sync_offset, 0, 1);
-                    lsc_atomic_update<atomic_op::inc, uint32_t, 1>(p_sync_buffer+pass_over_sync_offset, simd<uint32_t, 1>(0), 1);
-                }
-            }
-            // calculate each thread starting offset
-            barrier();
-            {
-                #pragma unroll
-                for (uint32_t s = 0; s<BIN_COUNT; s+=64) {
-                    bin_offset.template select<64, 1>(s).template bit_cast_view<uint32_t>() = lsc_slm_block_load<uint32_t, 64>(slm_incoming_start + s*sizeof(hist_t));
-                }
-                if (local_tid>0) {
-                    #pragma unroll
-                    for (uint32_t s = 0; s<BIN_COUNT; s+=64) {
-                        simd<hist_t, 64> group_local_sum;
-                        group_local_sum.template bit_cast_view<uint32_t>() = lsc_slm_block_load<uint32_t, 64>(slm_bin_hist_start + (local_tid-1)*HIST_STRIDE + s*sizeof(hist_t));
-                        bin_offset.template select<64, 1>(s) += group_local_sum;
-                    }
-                }
-                if ((local_tid > THREAD_PER_BIN_GROUP) && (local_tid % BIN_HEIGHT != 0)) {
-                    uint32_t prev_cum_rowid = (local_tid-1) / BIN_HEIGHT * BIN_HEIGHT - 1;
-                    #pragma unroll
-                    for (uint32_t s = 0; s<BIN_COUNT; s+=64) {
-                        simd<hist_t, 64> group_sum;
-                        group_sum.template bit_cast_view<uint32_t>() = lsc_slm_block_load<uint32_t, 64>(slm_bin_hist_start + prev_cum_rowid*HIST_STRIDE + s*sizeof(hist_t));
-                        bin_offset.template select<64, 1>(s) += group_sum;
-                    }
-                }
-            }
-        }
-    }
-    {
-        constexpr uint32_t BLOCK_SIZE = 32;
-        simd<device_addr_t, BLOCK_SIZE> write_addr;
-        simd<KeyT, BLOCK_SIZE> keys;
-        simd<bin_t, BLOCK_SIZE> bins;
-        // auto p_read = in + io_offset;
-        if (io_offset < n) {
-            for (uint32_t i=0; i<PROCESS_SIZE; i+=BLOCK_SIZE) {
-                simd<device_addr_t, BLOCK_SIZE> lane_id_block(0, 1);
-                simd_mask<BLOCK_SIZE> m = (io_offset+lane_id_block+i)<n;
-                simd<KeyT, BLOCK_SIZE> source = lsc_gather<KeyT, 1,
-                        // lsc_data_size::default_size, cache_hint::cached, cache_hint::cached, BLOCK_SIZE>(p_read+i, lane_id_block*uint32_t(sizeof(KeyT)), m);
-                        lsc_data_size::default_size, cache_hint::cached, cache_hint::cached, BLOCK_SIZE>(in, (io_offset + i + lane_id_block) * uint32_t(sizeof(KeyT)), m);
-                keys = merge(source, simd<KeyT, BLOCK_SIZE>(-1), m);
-
-                bins = (keys >> (stage * RADIX_BITS)) & MASK;
-                #pragma unroll
-                for (uint32_t s = 0; s<BLOCK_SIZE; s+=1) {
-                    write_addr[s] = bin_offset[bins[s]];
-                    bin_offset[bins[s]] += 1;
-                }
-                lsc_scatter<KeyT, 1, lsc_data_size::default_size, cache_hint::write_back, cache_hint::write_back, BLOCK_SIZE>(
-                    out,
-                    write_addr*sizeof(KeyT),
-                    keys, write_addr<n);
-            }
-        }
-    }
-}
-
-/* TODO:
-    Make sure the data is in rng after the last stage
-    Generate unique kernel names
-*/
-template <typename _Range, typename KeyT, uint32_t RADIX_BITS>
-void onesweep(sycl::queue &q, _Range&& rng, size_t n) {
-    using namespace sycl;
-    using namespace __ESIMD_NS;
-    using global_hist_t = uint32_t;
-
-    const uint32_t PROCESS_SIZE = 256;
-    constexpr uint32_t BINCOUNT = 1 << RADIX_BITS;
-    const uint32_t HW_TG_COUNT = 64;
-    constexpr uint32_t THREAD_PER_TG = 64;
-    uint32_t SWEEP_PROCESSING_SIZE = PROCESS_SIZE;
-    const uint32_t sweep_tg_count = div_up(n, THREAD_PER_TG*SWEEP_PROCESSING_SIZE);
-    const uint32_t sweep_threads = sweep_tg_count * THREAD_PER_TG;
-    constexpr uint32_t NBITS =  sizeof(KeyT) * 8;
-    constexpr uint32_t STAGES = div_up(NBITS, RADIX_BITS);
-
-    assert((SWEEP_PROCESSING_SIZE == 256) || (SWEEP_PROCESSING_SIZE == 512) || (SWEEP_PROCESSING_SIZE == 1024) || (SWEEP_PROCESSING_SIZE == 1536));
-
-    constexpr uint32_t SYNC_SEGMENT_COUNT = 64;
-    constexpr uint32_t SYNC_SEGMENT_SIZE_DW = 128;
-    constexpr uint32_t SYNC_BUFFER_SIZE = SYNC_SEGMENT_COUNT * SYNC_SEGMENT_SIZE_DW * sizeof(uint32_t); //bytes
-    constexpr uint32_t GLOBAL_OFFSET_SIZE = BINCOUNT * STAGES * sizeof(global_hist_t);
-    size_t temp_buffer_size = GLOBAL_OFFSET_SIZE + // global offset
-                              SYNC_BUFFER_SIZE;  // sync buffer
-
-    uint8_t *tmp_buffer = reinterpret_cast<uint8_t*>(sycl::malloc_device(temp_buffer_size, q));
-    auto p_global_offset = reinterpret_cast<uint32_t*>(tmp_buffer);
-    auto p_sync_buffer = reinterpret_cast<uint32_t*>(tmp_buffer + GLOBAL_OFFSET_SIZE);
-    auto p_output = sycl::malloc_device<KeyT>(n, q);
-
-    auto acc = rng.__m_acc;
-    using InT = decltype(acc);
-    using OutT = decltype(p_output);
-    {
-        nd_range<1> Range( (range<1>( HW_TG_COUNT * THREAD_PER_TG)), (range<1>(THREAD_PER_TG)) );
-        auto e = q.submit([&](handler &cgh) {
-            oneapi::dpl::__ranges::__require_access(cgh, rng);
-            cgh.parallel_for<class kernel_global_histogram>(
-                    Range, [=](nd_item<1> idx) [[intel::sycl_explicit_simd]] {
-                        global_histogram<InT, KeyT, RADIX_BITS, HW_TG_COUNT, THREAD_PER_TG> (idx, n, acc, p_global_offset, p_sync_buffer);
-                    });
-        });
-#if PRINT_HISTOGRAM
-        auto p_debug_histogram = sycl::malloc_host<uint32_t>(temp_buffer_size/sizeof(uint32_t), q);
-        q.copy<uint32_t>(p_global_offset, p_debug_histogram, temp_buffer_size/sizeof(uint32_t)).wait();
-        for(uint32_t i = 0; i < temp_buffer_size/sizeof(uint32_t); ++i)
-            std::cout << p_debug_histogram[i] << ' ';
-        std::cout << std::endl;
-#endif
-    }
-    {
-        auto e = q.submit([&](handler &cgh) {
-            cgh.parallel_for<class kernel_global_scan>(
-                nd_range<1>({STAGES * BINCOUNT}, {BINCOUNT}), [=](nd_item<1> idx) {
-                    uint32_t offset = idx.get_global_id(0);
-                    auto g = idx.get_group();
-                    uint32_t count = p_global_offset[offset];
-                    uint32_t presum = sycl::exclusive_scan_over_group(g, count, sycl::plus<KeyT>());
-                    p_global_offset[offset] = presum;
-                });
-            });
-    }
-    {
-        sycl::event e;
-        for (uint32_t stage = 0; stage < STAGES; stage++) {
-            nd_range<1> Range( (range<1>( sweep_tg_count * THREAD_PER_TG)), (range<1>(THREAD_PER_TG)) );
-            if (SWEEP_PROCESSING_SIZE == 256)
-            {
-                if((stage & 1) == 0)
-                {
-                    e = q.submit([&](handler &cgh) {
-                        oneapi::dpl::__ranges::__require_access(cgh, rng);
-                        cgh.parallel_for<class kernel_radix_sort_onesweep_256_0>(
-                            Range, [=](nd_item<1> idx) [[intel::sycl_explicit_simd]] {
-                                onesweep_kernel<InT, OutT, KeyT, RADIX_BITS, THREAD_PER_TG, 256> (idx, n, stage, acc, p_output, tmp_buffer);
-                            });
-                    });
-                }
-                else
-                {
-                    e = q.submit([&](handler &cgh) {
-                        oneapi::dpl::__ranges::__require_access(cgh, rng);
-                        cgh.parallel_for<class kernel_radix_sort_onesweep_256_1>(
-                            Range, [=](nd_item<1> idx) [[intel::sycl_explicit_simd]] {
-                                onesweep_kernel<OutT, InT, KeyT, RADIX_BITS, THREAD_PER_TG, 256> (idx, n, stage, p_output, acc, tmp_buffer);
-                            });
-                    });
-                }
-            }
-#if 0
-            else if (SWEEP_PROCESSING_SIZE == 512) {
-                e = q.submit([&](handler &cgh) {
-                    oneapi::dpl::__ranges::__require_access(cgh, rng);
-                    if((stage & 1) == 0)
-                    {
-                        cgh.parallel_for<class kernel_radix_sort_onesweep_512_0>(
-                                Range, [=](nd_item<1> idx) [[intel::sycl_explicit_simd]] {
-                                        onesweep_kernel<InT, OutT, KeyT, RADIX_BITS, THREAD_PER_TG, 512> (idx, n, stage, acc, p_output, tmp_buffer);
-                                });
-                    }
-                    else
-                    {
-                        cgh.parallel_for<class kernel_radix_sort_onesweep_512_1>(
-                                Range, [=](nd_item<1> idx) [[intel::sycl_explicit_simd]] {
-                                        onesweep_kernel<OutT, InT, KeyT, RADIX_BITS, THREAD_PER_TG, 512> (idx, n, stage, p_output, acc, tmp_buffer);
-                                });
-                    }
-                });
-            } else if (SWEEP_PROCESSING_SIZE == 1024) {
-                e = q.submit([&](handler &cgh) {
-                    oneapi::dpl::__ranges::__require_access(cgh, rng);
-                    if((stage & 1) == 0)
-                    {
-                        cgh.parallel_for<class kernel_radix_sort_onesweep_1024_0>(
-                                Range, [=](nd_item<1> idx) [[intel::sycl_explicit_simd]] {
-                                    onesweep_kernel<InT, OutT, KeyT, RADIX_BITS, THREAD_PER_TG, 1024> (idx, n, stage, acc, p_output, tmp_buffer);
-                                });
-                    }
-                    else
-                    {
-                        cgh.parallel_for<class kernel_radix_sort_onesweep_1024_1>(
-                                Range, [=](nd_item<1> idx) [[intel::sycl_explicit_simd]] {
-                                    onesweep_kernel<OutT, InT, KeyT, RADIX_BITS, THREAD_PER_TG, 1024> (idx, n, stage, p_output, acc, tmp_buffer);
-                                });
-                    }
-                });
-            } else if (SWEEP_PROCESSING_SIZE == 1536) {
-                e = q.submit([&](handler &cgh) {
-                    oneapi::dpl::__ranges::__require_access(cgh, rng);
-                    if((stage & 1) == 0)
-                    {
-                        cgh.parallel_for<class kernel_radix_sort_onesweep_1536_0>(
-                                Range, [=](nd_item<1> idx) [[intel::sycl_explicit_simd]] {
-                                    onesweep_kernel<InT, OutT, KeyT, RADIX_BITS, THREAD_PER_TG, 1536> (idx, n, stage, acc, p_output, tmp_buffer);
-                                });
-                    }
-                    else
-                    {
-                        cgh.parallel_for<class kernel_radix_sort_onesweep_1536_1>(
-                                Range, [=](nd_item<1> idx) [[intel::sycl_explicit_simd]] {
-                                    onesweep_kernel<OutT, InT, KeyT, RADIX_BITS, THREAD_PER_TG, 1536> (idx, n, stage, p_output, acc, tmp_buffer);
-                                });
-                    }
-                });
-            }
-#endif // 0
-        }
-        e.wait();
-    }
-    sycl::free(tmp_buffer, q);
-    sycl::free(p_output, q);
-}
-
-} // namespace oneapi::dpl::experimental::esimd::impl
-
+#include <type_traits>
 
 namespace oneapi::dpl::experimental::esimd
 {
@@ -964,35 +45,62 @@ namespace oneapi::dpl::experimental::esimd
          eSIMD operations:
             gather: Element type; can only be a 1,2,4-byte integer, sycl::half or float.
             lsc_gather: limited supported platforms: see https://intel.github.io/llvm-docs/doxygen/group__sycl__esimd__memory__lsc.html#ga250b3c0085f39c236582352fb711aadb)
-            one_workgroup:
-                RadixBits = 7 or 8: see BIN_WIDTH and scan implementation
 */
-// TODO: call it only for all_view (accessor) and guard_view (USM) ranges
-template <typename Policy, typename _Range, bool IsAscending = true, std::uint16_t WorkGroupSize = 256,
+// TODO: call it only for all_view (accessor) and guard_view (USM) ranges and probably subrange (according to the documentation, not tested)
+template <typename _ExecutionPolicy, typename _Range, bool IsAscending = true, std::uint16_t WorkGroupSize = 256,
           std::uint16_t ItemsPerWorkItem = 16, std::uint32_t RadixBits = 8>
 void
-radix_sort(Policy&& exec, _Range&& rng)
+radix_sort(_ExecutionPolicy&& __exec, _Range&& __rng)
 {
-    auto q = exec.queue();
-    using KeyT = oneapi::dpl::__internal::__value_t<_Range>;
+    using _KeyT = oneapi::dpl::__internal::__value_t<_Range>;
 
-    const ::std::size_t n = rng.size();
-#if 0
-    if (n <= 16384)
+    auto __q = __exec.queue();
+    const ::std::size_t __n = __rng.size();
+    auto __data = __rng.data();
+
+    if (__n <= 16384)
     {
-        oneapi::dpl::experimental::esimd::impl::one_workgroup<_Range, KeyT, RadixBits>(q, std::forward<_Range>(rng), n);
+        // TODO: allow passing accessor
+        // TODO: allow differnt sorting orders
+        // TODO: allow diferent types
+        // TODO: support c++17 (c++20 and newer are currently supported)
+        // TODO: generate unique kernel names
+        // TODO: allow all RadixBits values (only 7 or 8 are currently supported)
+        oneapi::dpl::experimental::esimd::impl::one_wg<_KeyT, decltype(__data), RadixBits>(__q, __data, __n);
     }
-    else if (n <= 262144)
+    else if (__n <= 262144)
     {
-        oneapi::dpl::experimental::esimd::impl::cooperative<_Range, KeyT, RadixBits>(q, std::forward<_Range>(rng), n);
+        // TODO: allow passing accessor
+        // TODO: allow differnt sorting orders
+        // TODO: allow diferent types
+        // TODO: support c++17 (c++20 and newer are currently supported)
+        // TODO: generate unique kernel names
+        oneapi::dpl::experimental::esimd::impl::cooperative<_KeyT, decltype(__data), RadixBits>(__q, __data, __n);
     }
-#endif // 0
-    // else
+    else
     {
-        // TODO: pass processing size basing on input size
-        oneapi::dpl::experimental::esimd::impl::onesweep<_Range, KeyT, RadixBits>(q, std::forward<_Range>(rng), n);
+        // TODO: allow passing accessor
+        // TODO: allow differnt sorting orders
+        // TODO: allow diferent types
+        // TODO: support c++17 (c++20 and newer are currently supported)
+        // TODO: generate unique kernel names
+        // TODO: avoid kernel duplication (generate the output storate with the same type as __data and use swap)
+        // TODO: allow different RadixBits, make sure the data is in __data after the last stage
+        // TODO: pass process_size according to __n
+        oneapi::dpl::experimental::esimd::impl::onesweep<_KeyT, decltype(__data), RadixBits>(__q, __data, __n, /*process_size*/ 512);
     }
 }
+
+template <typename _ExecutionPolicy, typename _Iterator, bool IsAscending = true, std::uint16_t WorkGroupSize = 256,
+          std::uint16_t ItemsPerWorkItem = 16, std::uint32_t RadixBits = 8>
+void
+radix_sort(_ExecutionPolicy&& __exec, _Iterator __first, _Iterator __last)
+{
+    auto __keep = oneapi::dpl::__ranges::__get_sycl_range<oneapi::dpl::__par_backend_hetero::access_mode::read_write, _Iterator>();
+    auto __rng = __keep(__first, __last);
+    radix_sort(::std::forward<_ExecutionPolicy>(__exec), __rng.all_view());
+}
+
 } // namespace oneapi::dpl::experimental::esimd
 
 #endif // _ONEDPL_parallel_backend_esimd_radix_sort_H
