+++ include/oneapi/dpl/pstl/utils_ranges.h
@@ -200,6 +200,12 @@ class guard_view
         return size() == 0;
     }
 
+    // Unified funciton to get pointer or accessor to use inside ESIMD kernels
+    _Iterator data()
+    {
+        return m_p;
+    }
+
   private:
     _Iterator m_p;     // a iterator (pointer)  to data in memory
     diff_type m_count; // size of the data
